fix[footer]: allow defining markup language
Custom footer is currently rendered through .RenderString, that defaults
to page's markup language, but page and footer could be written in two
different languages (eg. footer in Markdown and page in Asciidoc). In
this case the footer wouldn't be rendered correctly unless the markup
language used is specified. Now it is possible.